refactor: split analysis/thirdparty, analysis/clustering, and security/csp
Split three files approaching the 800 LOC limit:
- thirdparty.go (704) → thirdparty.go (455) + thirdparty_reputation.go (139) + thirdparty_summary.go (117)
- clustering.go (701) → clustering.go (427) + clustering_helpers.go (265)
- csp.go (666) → csp.go (379) + csp_helpers.go (246)
