+++ internal/analysis/clustering.go
@@ -1,6 +1,3 @@
-// Purpose: Owns clustering.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // clustering.go — Error clustering by root cause using 2-of-3 signal matching.
 // Groups related console errors by comparing normalized messages, application-level
 // stack frames (excluding framework code), and temporal proximity (<2s).
@@ -52,9 +49,6 @@ package analysis
 
 import (
 	"fmt"
-	"regexp"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -63,130 +57,6 @@ import (
 
 // Alert is an alias for types.Alert to avoid qualifying everywhere.
 type Alert = types.Alert
-
-// --- Stack Frame Parsing ---
-
-// StackFrame represents a parsed stack trace frame.
-type StackFrame struct {
-	Function    string
-	File        string
-	Line        int
-	Column      int
-	IsFramework bool
-}
-
-// Framework path patterns — frames from these paths are excluded from similarity comparison.
-var frameworkPatterns = []string{
-	"node_modules/react",
-	"node_modules/vue",
-	"node_modules/@angular",
-	"node_modules/svelte",
-	"webpack/bootstrap",
-	"webpack/runtime",
-	"zone.js",
-	"node_modules/rxjs",
-	"node_modules/core-js",
-}
-
-var (
-	// "    at FunctionName (file.js:line:col)"
-	stackFrameWithFunc = regexp.MustCompile(`^\s*at\s+(.+?)\s+\((.+?):(\d+):(\d+)\)`)
-	// "    at file.js:line:col"
-	stackFrameAnon = regexp.MustCompile(`^\s*at\s+(.+?):(\d+):(\d+)\s*$`)
-)
-
-// parseStackFrame parses a single stack trace line into a StackFrame.
-func parseStackFrame(line string) StackFrame {
-	line = strings.TrimSpace(line)
-
-	// Try "at Function (file:line:col)"
-	if m := stackFrameWithFunc.FindStringSubmatch(line); m != nil {
-		lineNum, _ := strconv.Atoi(m[3])
-		colNum, _ := strconv.Atoi(m[4])
-		frame := StackFrame{
-			Function: m[1],
-			File:     m[2],
-			Line:     lineNum,
-			Column:   colNum,
-		}
-		frame.IsFramework = isFrameworkPath(frame.File)
-		return frame
-	}
-
-	// Try "at file:line:col" (anonymous)
-	if m := stackFrameAnon.FindStringSubmatch(line); m != nil {
-		lineNum, _ := strconv.Atoi(m[2])
-		colNum, _ := strconv.Atoi(m[3])
-		frame := StackFrame{
-			Function: "<anonymous>",
-			File:     m[1],
-			Line:     lineNum,
-			Column:   colNum,
-		}
-		frame.IsFramework = isFrameworkPath(frame.File)
-		return frame
-	}
-
-	return StackFrame{Function: "<unknown>", File: line}
-}
-
-// isFrameworkPath returns true if the file path matches a known framework pattern.
-func isFrameworkPath(path string) bool {
-	for _, pattern := range frameworkPatterns {
-		if strings.Contains(path, pattern) {
-			return true
-		}
-	}
-	return false
-}
-
-// parseStack parses a multi-line stack trace into frames.
-func parseStack(stack string) []StackFrame {
-	if stack == "" {
-		return nil
-	}
-	lines := strings.Split(strings.TrimSpace(stack), "\n")
-	frames := make([]StackFrame, 0, len(lines))
-	for _, line := range lines {
-		if strings.TrimSpace(line) == "" {
-			continue
-		}
-		frames = append(frames, parseStackFrame(line))
-	}
-	return frames
-}
-
-// appFrames returns only non-framework frames.
-func appFrames(frames []StackFrame) []StackFrame {
-	result := make([]StackFrame, 0)
-	for _, f := range frames {
-		if !f.IsFramework {
-			result = append(result, f)
-		}
-	}
-	return result
-}
-
-// --- Message Normalization ---
-
-var (
-	clusterUUIDRegex      = regexp.MustCompile(`[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}`)
-	clusterURLRegex       = regexp.MustCompile(`https?://[^\s"']+`)
-	clusterTimestampRegex = regexp.MustCompile(`\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}[^\s]*`)
-	clusterNumericIDRegex = regexp.MustCompile(`\b\d{3,}\b`) // 3+ digit numbers as IDs
-)
-
-// normalizeErrorMessage replaces variable content with placeholders.
-func normalizeErrorMessage(msg string) string {
-	// Order matters: UUIDs before numeric IDs (UUIDs contain digits)
-	result := clusterUUIDRegex.ReplaceAllString(msg, "{uuid}")
-	result = clusterURLRegex.ReplaceAllString(result, "{url}")
-	result = clusterTimestampRegex.ReplaceAllString(result, "{timestamp}")
-	result = clusterNumericIDRegex.ReplaceAllString(result, "{id}")
-	return result
-}
-
-// --- Error Instance ---
 
 // ErrorInstance represents a single error received from the extension.
 type ErrorInstance struct {
@@ -197,8 +67,6 @@ type ErrorInstance struct {
 	ErrorType string // TypeError, ReferenceError, etc.
 	Severity  string // error, warning
 }
-
-// --- Cluster ---
 
 // ErrorCluster groups related errors by root cause.
 type ErrorCluster struct {
@@ -216,8 +84,6 @@ type ErrorCluster struct {
 	alertedAt3     bool // track if we already alerted for this cluster
 }
 
-// --- Cluster Manager ---
-
 // ClusterManager manages error clustering with session-scoped lifecycle.
 type ClusterManager struct {
 	mu             sync.RWMutex
@@ -227,6 +93,26 @@ type ClusterManager struct {
 	expiryDuration time.Duration
 	pendingAlert   *Alert
 	totalErrors    int
+}
+
+// ClusterAnalysisResponse is the response for analyze(target: "errors").
+type ClusterAnalysisResponse struct {
+	Clusters          []ClusterSummary `json:"clusters"`
+	UnclusteredErrors int              `json:"unclustered_errors"`
+	TotalErrors       int              `json:"total_errors"`
+	Summary           string           `json:"summary"`
+}
+
+// ClusterSummary is a single cluster in the analysis response.
+type ClusterSummary struct {
+	ID                string   `json:"id"`
+	RepresentativeMsg string   `json:"representative_error"`
+	RootCause         string   `json:"root_cause"`
+	InstanceCount     int      `json:"instance_count"`
+	FirstSeen         string   `json:"first_seen"`
+	LastSeen          string   `json:"last_seen"`
+	AffectedFiles     []string `json:"affected_components"`
+	Severity          string   `json:"severity"`
 }
 
 // NewClusterManager creates an empty cluster manager.
@@ -335,94 +221,43 @@ func (cm *ClusterManager) addUnclustered(err ErrorInstance) {
 	}
 }
 
-// matchesCluster checks if an error matches an existing cluster.
-func (cm *ClusterManager) matchesCluster(cluster *ErrorCluster, err ErrorInstance, appFr []StackFrame, normMsg string) bool {
-	// For errors without stacks, message match alone is sufficient
-	if err.Stack == "" && len(cluster.Instances) > 0 && cluster.Instances[0].Stack == "" && cluster.NormalizedMsg == normMsg {
-		return true
-	}
-	return cm.clusterSignalCount(cluster, err, appFr, normMsg) >= 2
-}
+// addToCluster adds an error to an existing cluster.
+func (cm *ClusterManager) addToCluster(cluster *ErrorCluster, err ErrorInstance) {
+	cluster.InstanceCount++
+	cluster.LastSeen = err.Timestamp
 
-// clusterSignalCount counts how many of the 3 signals match between an error and a cluster.
-func (cm *ClusterManager) clusterSignalCount(cluster *ErrorCluster, err ErrorInstance, appFr []StackFrame, normMsg string) int {
-	signals := 0
-	if cluster.NormalizedMsg == normMsg {
-		signals++
-	}
-	if len(appFr) > 0 && len(cluster.CommonFrames) > 0 && countSharedFrames(appFr, cluster.CommonFrames) >= 1 {
-		signals++
-	}
-	if !cluster.LastSeen.IsZero() && err.Timestamp.Sub(cluster.LastSeen) < 2*time.Second {
-		signals++
-	}
-	return signals
-}
-
-// countSignals counts how many signals match between two errors.
-// Core signal matching logic used by matchesCluster decision.
-// Returns count of matched signals (0-3).
-//
-// Three Signals Evaluated:
-//  1. Message Signal: normalized messages identical
-//  2. Frames Signal: 2+ shared application-level frames (countSharedFrames)
-//  3. Temporal Signal: time between errors < 2 seconds (temporal proximity)
-//
-// Decision Rule (caller's responsibility):
-//   - 2+ signals matching → likely same root cause
-//   - 1 signal matching → could be coincidence, require more evidence
-//   - 0 signals matching → different errors
-//
-// Note on Signal Independence:
-//   - Signals are roughly independent (one signal doesn't imply another)
-//   - Message alone could be coincidence (same generic message, different code)
-//   - Frame signature alone could be coincidence (different bugs in same function)
-//   - Time proximity alone could be coincidence (rapid error bursts from unrelated bugs)
-//   - 2+ signals together indicate correlation likely > coincidence
-//
-// Caller must provide:
-//   - existing: Error already in this cluster
-//   - new: Error being considered for addition
-//   - newAppFr: Application frames from new error (already filtered via appFrames())
-//   - newNormMsg: Normalized message from new error (already normalized)
-func (cm *ClusterManager) countSignals(existing, new ErrorInstance, newAppFr []StackFrame, newNormMsg string) int {
-	signals := 0
-
-	// Message similarity
-	existingNorm := normalizeErrorMessage(existing.Message)
-	if existingNorm == newNormMsg {
-		signals++
+	// Cap instances at 20
+	if len(cluster.Instances) < 20 {
+		cluster.Instances = append(cluster.Instances, err)
 	}
 
-	// Stack similarity
-	existingFrames := appFrames(parseStack(existing.Stack))
-	if len(existingFrames) > 0 && len(newAppFr) > 0 {
-		if countSharedFrames(existingFrames, newAppFr) >= 1 {
-			signals++
+	// Update affected files
+	frames := appFrames(parseStack(err.Stack))
+	for _, f := range frames {
+		found := false
+		for _, existing := range cluster.AffectedFiles {
+			if existing == f.File {
+				found = true
+				break
+			}
+		}
+		if !found {
+			cluster.AffectedFiles = append(cluster.AffectedFiles, f.File)
 		}
 	}
 
-	// Temporal proximity
-	if new.Timestamp.Sub(existing.Timestamp) < 2*time.Second {
-		signals++
-	}
-
-	return signals
-}
-
-// countSharedFrames counts frames that appear in both slices (by file:line).
-func countSharedFrames(a, b []StackFrame) int {
-	bSet := make(map[string]bool)
-	for _, f := range b {
-		bSet[fmt.Sprintf("%s:%d", f.File, f.Line)] = true
-	}
-	count := 0
-	for _, f := range a {
-		if bSet[fmt.Sprintf("%s:%d", f.File, f.Line)] {
-			count++
+	// Alert at 3 instances (once per cluster)
+	if cluster.InstanceCount == 3 && !cluster.alertedAt3 {
+		cluster.alertedAt3 = true
+		cm.pendingAlert = &Alert{
+			Severity:  "error",
+			Category:  "error_cluster",
+			Title:     fmt.Sprintf("Error cluster: %d related errors from %s", cluster.InstanceCount, cluster.RootCause),
+			Detail:    fmt.Sprintf("Cluster %s: %s", cluster.ID, cluster.Representative.Message),
+			Timestamp: time.Now().UTC().Format(time.RFC3339),
+			Source:    "clustering",
 		}
 	}
-	return count
 }
 
 // createCluster creates a new cluster from two matching errors.
@@ -486,93 +321,6 @@ func (cm *ClusterManager) createCluster(first, second ErrorInstance, normMsg str
 		Severity:       severity,
 	}
 }
-
-// addToCluster adds an error to an existing cluster.
-func (cm *ClusterManager) addToCluster(cluster *ErrorCluster, err ErrorInstance) {
-	cluster.InstanceCount++
-	cluster.LastSeen = err.Timestamp
-
-	// Cap instances at 20
-	if len(cluster.Instances) < 20 {
-		cluster.Instances = append(cluster.Instances, err)
-	}
-
-	// Update affected files
-	frames := appFrames(parseStack(err.Stack))
-	for _, f := range frames {
-		found := false
-		for _, existing := range cluster.AffectedFiles {
-			if existing == f.File {
-				found = true
-				break
-			}
-		}
-		if !found {
-			cluster.AffectedFiles = append(cluster.AffectedFiles, f.File)
-		}
-	}
-
-	// Alert at 3 instances (once per cluster)
-	if cluster.InstanceCount == 3 && !cluster.alertedAt3 {
-		cluster.alertedAt3 = true
-		cm.pendingAlert = &Alert{
-			Severity:  "error",
-			Category:  "error_cluster",
-			Title:     fmt.Sprintf("Error cluster: %d related errors from %s", cluster.InstanceCount, cluster.RootCause),
-			Detail:    fmt.Sprintf("Cluster %s: %s", cluster.ID, cluster.Representative.Message),
-			Timestamp: time.Now().UTC().Format(time.RFC3339),
-			Source:    "clustering",
-		}
-	}
-}
-
-// findCommonFrames returns frames present in both slices.
-func findCommonFrames(a, b []StackFrame) []StackFrame {
-	bSet := make(map[string]StackFrame)
-	for _, f := range b {
-		key := fmt.Sprintf("%s:%d", f.File, f.Line)
-		bSet[key] = f
-	}
-	var common []StackFrame
-	for _, f := range a {
-		key := fmt.Sprintf("%s:%d", f.File, f.Line)
-		if _, ok := bSet[key]; ok {
-			common = append(common, f)
-		}
-	}
-	return common
-}
-
-// inferRootCause returns the deepest common app-code frame, or the normalized message.
-func inferRootCause(commonFrames []StackFrame, normMsg string) string {
-	// The deepest frame (first in the list, since stacks go from deepest to shallowest)
-	for _, f := range commonFrames {
-		if !f.IsFramework {
-			if f.Function != "<anonymous>" && f.Function != "<unknown>" {
-				return fmt.Sprintf("%s (%s:%d)", f.Function, f.File, f.Line)
-			}
-			return fmt.Sprintf("%s:%d", f.File, f.Line)
-		}
-	}
-	return normMsg
-}
-
-// collectAffectedFiles returns unique source files from two frame sets.
-func collectAffectedFiles(a, b []StackFrame) []string {
-	seen := make(map[string]bool)
-	var files []string
-	for _, frames := range [][]StackFrame{a, b} {
-		for _, f := range frames {
-			if f.File != "" && !f.IsFramework && !seen[f.File] {
-				seen[f.File] = true
-				files = append(files, f.File)
-			}
-		}
-	}
-	return files
-}
-
-// --- Query Methods ---
 
 // GetClusters returns all active clusters.
 func (cm *ClusterManager) GetClusters() []*ErrorCluster {
@@ -641,28 +389,6 @@ func (cm *ClusterManager) Cleanup() {
 		}
 	}
 	cm.clusters = active
-}
-
-// --- Analysis Response ---
-
-// ClusterAnalysisResponse is the response for analyze(target: "errors").
-type ClusterAnalysisResponse struct {
-	Clusters          []ClusterSummary `json:"clusters"`
-	UnclusteredErrors int              `json:"unclustered_errors"`
-	TotalErrors       int              `json:"total_errors"`
-	Summary           string           `json:"summary"`
-}
-
-// ClusterSummary is a single cluster in the analysis response.
-type ClusterSummary struct {
-	ID                string   `json:"id"`
-	RepresentativeMsg string   `json:"representative_error"`
-	RootCause         string   `json:"root_cause"`
-	InstanceCount     int      `json:"instance_count"`
-	FirstSeen         string   `json:"first_seen"`
-	LastSeen          string   `json:"last_seen"`
-	AffectedFiles     []string `json:"affected_components"`
-	Severity          string   `json:"severity"`
 }
 
 // GetAnalysisResponse builds the response for analyze(target: "errors").
